+++ accounts/keystore/account_cache.go
@@ -1,25 +1,16 @@
 package keystore
 
 import (
-	"bufio"
-	"encoding/json"
+	"errors"
 	"os"
-	"path/filepath"
-	"slices"
-	"sort"
-	"strings"
+	"path"
 	"sync"
 	"time"
 
 	"github.com/0xPolygon/polygon-edge/accounts"
 	"github.com/0xPolygon/polygon-edge/types"
-	mapset "github.com/deckarep/golang-set/v2"
 	"github.com/hashicorp/go-hclog"
 )
-
-func byURL(a, b accounts.Account) int {
-	return a.URL.Cmp(b.URL)
-}
 
 // KeyStoreScheme is the protocol scheme prefixing account and wallet URLs.
 const KeyStoreScheme = "keystore"
@@ -29,10 +20,8 @@ const minReloadInterval = 2 * time.Second
 type accountCache struct {
 	logger   hclog.Logger
 	keydir   string
-	watcher  *watcher
 	mu       sync.Mutex
-	all      []accounts.Account
-	byAddr   map[types.Address][]accounts.Account
+	allMap   map[types.Address]encryptedKeyJSONV3
 	throttle *time.Timer
 	notify   chan struct{}
 	fileC    fileCache
@@ -42,49 +31,60 @@ func newAccountCache(keyDir string, logger hclog.Logger) (*accountCache, chan st
 	ac := &accountCache{
 		logger: logger,
 		keydir: keyDir,
-		byAddr: make(map[types.Address][]accounts.Account),
 		notify: make(chan struct{}, 1),
-		fileC:  fileCache{all: mapset.NewThreadUnsafeSet[string]()},
 	}
-	ac.watcher = newWatcher(ac, logger)
+
+	keyPath := path.Join(keyDir, "keys.txt")
+
+	ac.fileC = fileCache{all: make(map[types.Address]encryptedKeyJSONV3), keyDir: keyPath}
+
+	if _, err := os.Stat(keyPath); errors.Is(err, os.ErrNotExist) {
+		os.Create(keyDir)
+	}
+
+	ac.scanAccounts()
 
 	return ac, ac.notify
 }
 
 func (ac *accountCache) accounts() []accounts.Account {
-	ac.maybeReload()
+	ac.scanAccounts()
+
 	ac.mu.Lock()
 	defer ac.mu.Unlock()
 
-	cpy := make([]accounts.Account, len(ac.all))
-	copy(cpy, ac.all)
+	cpy := make([]accounts.Account, len(ac.allMap))
+	for addr := range ac.allMap {
+		cpy = append(cpy, accounts.Account{Address: addr})
+	}
 
 	return cpy
 }
 
 func (ac *accountCache) hasAddress(addr types.Address) bool {
-	ac.maybeReload()
+	ac.scanAccounts()
 
 	ac.mu.Lock()
 	defer ac.mu.Unlock()
 
-	return len(ac.byAddr[addr]) > 0
+	_, ok := ac.allMap[addr]
+
+	return ok
 }
 
 func (ac *accountCache) add(newAccount accounts.Account) {
+	ac.scanAccounts()
+
 	ac.mu.Lock()
 	defer ac.mu.Unlock()
 
-	i := sort.Search(len(ac.all), func(i int) bool { return ac.all[i].URL.Cmp(newAccount.URL) >= 0 })
-	if i < len(ac.all) && ac.all[i] == newAccount {
+	if _, ok := ac.allMap[newAccount.Address]; ok {
 		return
 	}
 
-	// newAccount is not in the cache.
-	ac.all = append(ac.all, accounts.Account{})
-	copy(ac.all[i+1:], ac.all[i:])
-	ac.all[i] = newAccount
-	ac.byAddr[newAccount.Address] = append(ac.byAddr[newAccount.Address], newAccount)
+	ac.allMap[newAccount.Address] = encryptedKeyJSONV3{} // TO DO
+
+	ac.fileC.saveData(ac.allMap)
 }
 
 // note: removed needs to be unique here (i.e. both File and Address must be set).
@@ -92,127 +92,31 @@ func (ac *accountCache) delete(removed accounts.Account) {
 	ac.mu.Lock()
 	defer ac.mu.Unlock()
 
-	ac.all = removeAccount(ac.all, removed)
-	if ba := removeAccount(ac.byAddr[removed.Address], removed); len(ba) == 0 {
-		delete(ac.byAddr, removed.Address)
-	} else {
-		ac.byAddr[removed.Address] = ba
-	}
-}
+	ac.scanAccounts()
 
-// deleteByFile removes an account referenced by the given path.
-func (ac *accountCache) deleteByFile(path string) {
-	ac.mu.Lock()
-	defer ac.mu.Unlock()
+	delete(ac.allMap, removed.Address)
 
-	i := sort.Search(len(ac.all), func(i int) bool { return ac.all[i].URL.Path >= path })
-
-	if i < len(ac.all) && ac.all[i].URL.Path == path {
-		removed := ac.all[i]
-		ac.all = append(ac.all[:i], ac.all[i+1:]...)
-
-		if ba := removeAccount(ac.byAddr[removed.Address], removed); len(ba) == 0 {
-			delete(ac.byAddr, removed.Address)
-		} else {
-			ac.byAddr[removed.Address] = ba
-		}
-	}
-}
-
-// watcherStarted returns true if the watcher loop started running (even if it
-// has since also ended).
-func (ac *accountCache) watcherStarted() bool {
-	ac.mu.Lock()
-
-	defer ac.mu.Unlock()
-
-	return ac.watcher.running || ac.watcher.runEnded
-}
-
-func removeAccount(slice []accounts.Account, elem accounts.Account) []accounts.Account {
-	for i := range slice {
-		if slice[i] == elem {
-			return append(slice[:i], slice[i+1:]...)
-		}
-	}
-
-	return slice
+	ac.fileC.saveData(ac.allMap)
 }
 
 // find returns the cached account for address if there is a unique match.
 // The exact matching rules are explained by the documentation of accounts.Account.
 // Callers must hold ac.mu.
 func (ac *accountCache) find(a accounts.Account) (accounts.Account, error) {
-	// Limit search to address candidates if possible.
-	matches := ac.all
-	if (a.Address != types.Address{}) {
-		matches = ac.byAddr[a.Address]
-	}
+	ac.scanAccounts()
 
-	if a.URL.Path != "" {
-		// If only the basename is specified, complete the path.
-		if !strings.ContainsRune(a.URL.Path, filepath.Separator) {
-			a.URL.Path = filepath.Join(ac.keydir, a.URL.Path)
-		}
-
-		for i := range matches {
-			if matches[i].URL == a.URL {
-				return matches[i], nil
-			}
-		}
-
-		if (a.Address == types.Address{}) {
-			return accounts.Account{}, accounts.ErrNoMatch
-		}
-	}
-
-	switch len(matches) {
-	case 1:
-		return matches[0], nil
-	case 0:
-		return accounts.Account{}, accounts.ErrNoMatch
-	default:
-		err := &accounts.AmbiguousAddrError{Addr: a.Address, Matches: make([]accounts.Account, len(matches))}
-		copy(err.Matches, matches)
-		slices.SortFunc(err.Matches, byURL)
-
-		return accounts.Account{}, err
-	}
-}
-
-func (ac *accountCache) maybeReload() {
 	ac.mu.Lock()
+	defer ac.mu.Unlock()
 
-	if ac.watcher.running {
-		ac.mu.Unlock()
-
-		return // A watcher is running and will keep the cache up-to-date.
+	if _, ok := ac.allMap[a.Address]; ok {
+		return a, nil
 	}
 
-	if ac.throttle == nil {
-		ac.throttle = time.NewTimer(0)
-	} else {
-		select {
-		case <-ac.throttle.C:
-		default:
-			ac.mu.Unlock()
-
-			return // The cache was reloaded recently.
-		}
-	}
-	// No watcher running, start it.
-	ac.watcher.start()
-	ac.throttle.Reset(minReloadInterval)
-	ac.mu.Unlock()
-
-	if err := ac.scanAccounts(); err != nil {
-		ac.logger.Info("reload", "failed to scan accounts", err)
-	}
+	return accounts.Account{}, accounts.ErrNoMatch
 }
 
 func (ac *accountCache) close() {
 	ac.mu.Lock()
-	ac.watcher.close()
 
 	if ac.throttle != nil {
 		ac.throttle.Stop()
@@ -230,82 +134,26 @@ func (ac *accountCache) close() {
 // updates the account cache accordingly
 func (ac *accountCache) scanAccounts() error {
 	// Scan the entire folder metadata for file changes
-	creates, deletes, updates, err := ac.fileC.scan(ac.keydir)
+
+	ac.mu.Lock()
+	defer ac.mu.Unlock()
+
+	accs, err := ac.fileC.scanOneFile()
 	if err != nil {
 		ac.logger.Debug("Failed to reload keystore contents", "err", err)
 
 		return err
 	}
 
-	if creates.Cardinality() == 0 && deletes.Cardinality() == 0 && updates.Cardinality() == 0 {
-		return nil
-	}
-	// Create a helper method to scan the contents of the key files
-	var (
-		buf = new(bufio.Reader)
-		key struct {
-			Address string `json:"address"`
-		}
-	)
+	ac.allMap = make(map[types.Address]encryptedKeyJSONV3)
 
-	readAccount := func(path string) *accounts.Account {
-		fd, err := os.Open(path)
-		if err != nil {
-			ac.logger.Trace("Failed to open keystore file", "path", path, "err", err)
-
-			return nil
-		}
-
-		defer fd.Close()
-
-		buf.Reset(fd)
-		// Parse the address.
-		key.Address = ""
-		err = json.NewDecoder(buf).Decode(&key)
-		addr := types.StringToAddress(key.Address)
-
-		switch {
-		case err != nil:
-			ac.logger.Debug("Failed to decode keystore key", "path", path, "err", err)
-		case addr == types.Address{}:
-			ac.logger.Debug("Failed to decode keystore key", "path", path, "err", "missing or zero address")
-		default:
-			return &accounts.Account{
-				Address: addr,
-				URL:     accounts.URL{Scheme: KeyStoreScheme, Path: path},
-			}
-		}
-
-		return nil
-	}
-	// Process all the file diffs
-	start := time.Now().UTC()
-
-	for _, path := range creates.ToSlice() {
-		if a := readAccount(path); a != nil {
-			ac.add(*a)
-		}
+	for addr, key := range accs {
+		ac.allMap[addr] = key
 	}
 
-	for _, path := range deletes.ToSlice() {
-		ac.deleteByFile(path)
-	}
+	ac.notify <- struct{}{}
 
-	for _, path := range updates.ToSlice() {
-		ac.deleteByFile(path)
-
-		if a := readAccount(path); a != nil {
-			ac.add(*a)
-		}
-	}
-
-	end := time.Now().UTC()
-
-	select {
-	case ac.notify <- struct{}{}:
-	default:
-	}
-	ac.logger.Trace("Handled keystore changes", "time", end.Sub(start))
+	ac.logger.Trace("Handled keystore changes")
 
 	return nil
 }
